# Optimizations for extract and format