Switches to working with bytes (`u8`s), which is slightly more efficient than `chars` due to variable length char-codes of utf8 not having to be supported/validated. Also uses `simd` via `memchr` for byte searching and removes an allocation in `format` by handing over control of the `Vec<u8>` parts directly to the `String`.
```
cargo bench
   Compiling overpunch v0.3.0 (/Users/josephmoniz/src/overpunch)
    Finished `bench` profile [optimized] target(s) in 0.88s
     Running unittests src/lib.rs (target/release/deps/overpunch-244dc83774a84c8a)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running benches/benchmark.rs (target/release/deps/benchmark-c33f02f3f256d33c)
Gnuplot not found, using plotters backend
convert_from_signed_format
                        time:   [12.155 ns 12.252 ns 12.353 ns]
                        change: [-15.112% -14.264% -13.374%] (p = 0.00 < 0.05)
                        Performance has improved.

convert_to_signed_format
                        time:   [30.219 ns 30.256 ns 30.298 ns]
                        change: [-43.997% -43.654% -43.333%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe

extract                 time:   [8.0216 ns 8.0296 ns 8.0390 ns]
                        change: [-18.286% -17.977% -17.666%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe

format                  time:   [28.737 ns 29.002 ns 29.279 ns]
                        change: [-44.112% -43.656% -43.189%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  10 (10.00%) high mild
  9 (9.00%) high severe
  ```